updated put method and add character

+++ Services/CharacterService/CharacterService.cs
@@ -41,6 +41,7 @@ namespace dotnet_Rpg6.Services.CharacterService
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
             .Select(c => _mapper.Map<GetCharacterDto>(c))
             .ToListAsync();
             return serviceResponse;  
@@ -52,12 +53,21 @@ namespace dotnet_Rpg6.Services.CharacterService
             
             try
             {
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
+            Character character = await _context.Characters
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+            if(character != null)
+            {
             _context.Characters.Remove(character);  
             await _context.SaveChangesAsync(); 
-
-
-            response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList(); 
+            response.Data = _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList(); 
+            }
+            else 
+            {
+                response.Success = false;
+                response.Message = "Character not found";
+            }
             } 
             catch(Exception ex)
             {
@@ -81,7 +91,8 @@ namespace dotnet_Rpg6.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -93,7 +104,11 @@ namespace dotnet_Rpg6.Services.CharacterService
             try
             {
             var character = await _context.Characters
+            .Include(c => c.User)
             .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+
+            if(character.User.Id == GetUserId())
+            { 
 
             _mapper.Map(updatedCharacter, character);
             character.Name = updatedCharacter.Name;
@@ -107,6 +122,12 @@ namespace dotnet_Rpg6.Services.CharacterService
               
             response.Data = _mapper.Map<GetCharacterDto>(character);
             } 
+            else 
+            {
+                response.Success = false;
+                response.Message = "Character not found.";
+            }
+            }
             catch(Exception ex)
             {
                response.Success = false;
